properties in templates
