+++ src/RefDocGen/TemplateModels/Builders/CSharpTemplateModelBuilder.cs
@@ -7,15 +7,22 @@ internal class CSharpTemplateModelBuilder : ITemplateModelBuilder
     public ClassTemplateModel CreateClassTemplateModel(ClassIntermed classIntermed)
     {
         var fields = classIntermed.Fields.Select(CreateFieldTemplateModels).ToArray();
+        var properties = classIntermed.Properties.Select(CreatePropertyTemplateModels).ToArray();
         var methods = classIntermed.Methods.Select(CreateMethodTemplateModels).ToArray();
 
-        return new ClassTemplateModel(classIntermed.Name, string.Empty, [classIntermed.AccessibilityModifier.ToString()], fields, methods);
+        return new ClassTemplateModel(classIntermed.Name, string.Empty, [classIntermed.AccessibilityModifier.ToString()], fields, properties, methods);
     }
 
     private FieldTemplateModel CreateFieldTemplateModels(FieldIntermed fieldIntermed)
     {
         string[] modifiers = [fieldIntermed.AccessibilityModifier.ToString()];
         return new FieldTemplateModel(fieldIntermed.Name, fieldIntermed.Type, string.Empty, modifiers);
+    }
+
+    private PropertyTemplateModel CreatePropertyTemplateModels(PropertyIntermed propertyIntermed)
+    {
+        string[] modifiers = [propertyIntermed.AccessibilityModifier.ToString()];
+        return new PropertyTemplateModel(propertyIntermed.Name, propertyIntermed.Type, string.Empty, modifiers, propertyIntermed.Getter is not null, propertyIntermed.Setter is not null, [], []);
     }
 
     private MethodTemplateModel CreateMethodTemplateModels(MethodIntermed methodIntermed)
