+++ src/RefDocGen/TemplateModels/ClassTemplateModel.cs
@@ -1,5 +1,5 @@
 namespace RefDocGen.TemplateModels;
 
-public record ClassTemplateModel(string Name, string DocComment, string[] Modifiers, FieldTemplateModel[] Fields, MethodTemplateModel[] Methods)
+public record ClassTemplateModel(string Name, string DocComment, string[] Modifiers, FieldTemplateModel[] Fields, PropertyTemplateModel[] Properties, MethodTemplateModel[] Methods)
 {
 }
